Refactoring Project Structure
So I'll try to help a bit by making some changes in folder structure to make this project little bit easier to maintain. Let me know if it has any issues
and sorry for my bad english :D
